+++ src/knowledge_ingestor/extractors/__init__.py
@@ -1,1 +1,3 @@
-# TODO
+from .article import ExtractedArticle, extract_article
+
+__all__ = ["ExtractedArticle", "extract_article"]